Movie componenet modifications
