+++ src/modules/movies/Movie.jsx
@@ -1,21 +1,14 @@
-import { useDispatch, useSelector } from "react-redux";
-import starredSlice from "../../data/starredSlice";
-import watchLaterSlice from "../../data/watchLaterSlice";
 import placeholder from "../../assets/not-found-500X750.jpeg";
 import YoutubePlayer from "../../components/YoutubePlayer";
 import { useState } from "react";
 import { API_KEY, ENDPOINT, MOVIE_POSTER_DEFAULT } from "../../constants";
+import MovieStarButton from "./MovieStarButton";
+import MovieWatchLaterButton from "./MovieWatchLaterButton";
 
 const Movie = ({ movie }) => {
-  const state = useSelector((state) => state);
-  const { starred, watchLater } = state;
-  const { starMovie, unstarMovie } = starredSlice.actions;
-  const { addToWatchLater, removeFromWatchLater } = watchLaterSlice.actions;
   const [trailerData, setTrailerData] = useState(null);
   const [isOpen, setOpen] = useState(false);
   const [isCardOpened, setIsCardOpened] = useState(false);
-
-  const dispatch = useDispatch();
 
   const closeModal = () => setOpen(false);
 
@@ -50,87 +43,6 @@ const Movie = ({ movie }) => {
     }
   };
 
-  const StarButton = () => {
-    const isStarred = starred.starredMovies.some(
-      (starredMovie) => starredMovie.id === movie.id
-    );
-    const handleAddToStar = () => {
-      dispatch(
-        starMovie({
-          id: movie.id,
-          overview: movie.overview,
-          release_date: movie.release_date?.substring(0, 4),
-          poster_path: movie.poster_path,
-          title: movie.title,
-        })
-      );
-    };
-
-    const handleRemoveFromStar = () => {
-      dispatch(unstarMovie(movie));
-    };
-
-    return isStarred ? (
-      <span
-        className="btn-star"
-        data-testid="unstar-link"
-        onClick={handleRemoveFromStar}
-      >
-        <i className="bi bi-star-fill" data-testid="star-fill" />
-      </span>
-    ) : (
-      <span
-        className="btn-star"
-        data-testid="starred-link"
-        onClick={handleAddToStar}
-      >
-        <i className="bi bi-star" />
-      </span>
-    );
-  };
-
-  const WatchLaterButton = () => {
-    const isMovieInWatchLater = watchLater.watchLaterMovies.some(
-      (watchLaterMovie) => watchLaterMovie.id === movie.id
-    );
-
-    const handleAddToWatchLater = () => {
-      dispatch(
-        addToWatchLater({
-          id: movie.id,
-          overview: movie.overview,
-          release_date: movie.release_date?.substring(0, 4),
-          poster_path: movie.poster_path,
-          title: movie.title,
-        })
-      );
-    };
-
-    const handleRemoveFromWatchLater = () => {
-      dispatch(removeFromWatchLater(movie));
-    };
-
-    return isMovieInWatchLater ? (
-      <button
-        type="button"
-        data-testid="remove-watch-later"
-        className="btn btn-light btn-watch-later blue"
-        onClick={handleRemoveFromWatchLater}
-      >
-        <i className="bi bi-check"></i>
-      </button>
-    ) : (
-      <button
-        type="button"
-        data-testid="watch-later"
-        className="btn btn-light btn-watch-later"
-        onClick={handleAddToWatchLater}
-      >
-        Watch Later
-      </button>
-    );
-  };
-
   return (
     <>
       {isOpen ? (
@@ -151,8 +63,8 @@ const Movie = ({ movie }) => {
             <div className="info_panel">
               <div className="overview">{movie.overview}</div>
               <div className="year">{movie.release_date?.substring(0, 4)}</div>
-              <StarButton />
-              <WatchLaterButton />
+              <MovieStarButton movie={movie} />
+              <MovieWatchLaterButton movie={movie} />
               <button
                 type="button"
                 className="btn btn-dark"
